refactor(LineAcceptanceTest) 안쓰는 메서드나 중복 메서드 정리
